Remake project in homepage Correction to project page
