+++ app/(client)/projects/styles.css
@@ -1,0 +1,13 @@
+.projects {
+  width: 100%;
+  margin-inline: 10px;
+  display: grid;
+  grid-template-columns: 50% 50%;
+  grid-template-rows: auto;
+  row-gap: 2rem;
+}
+
+.projects .card {
+  display: flex;
+  justify-content: center;
+}
