+++ app/globals.css
@@ -1,3 +1,11 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
+
+.header {
+  margin-bottom: 1rem;
+  font-size: 2.25rem;
+  line-height: 2.5rem;
+  color: rgb(30 64 175 / var(--tw-text-opacity, 1));
+  --tw-text-opacity: 1;
+}
