+++ app/types/components.ts
@@ -3,3 +3,9 @@ export interface CardProps {
   description: string
   imageUrl?: string
 }
+
+export interface ProjectProps {
+  id: number
+  name: string
+  description: string
+}
